Changed floating time signature context spacing slightly.
The files created by ...

    lilypondfiletools.make_floating_time_signature_lilypond_file()

... now set spacing of the floating time signature context to 10 vertical units
instead of 12 vertical units. This draws the floating time signature context
closer to the fully written-out notation below it and leads to a cleaner (and
very slightly tighter) look in the doc examples that use this function.

In the mainline it is only the rhythm-maker doc examples that use this
function. The change will become visible the next time you do a clean rebuild
of the mainline docs.

All of the musicexpressiontools collateral in experimental use the function.
The change will only become visible when we decide to do a batch update of all
collateral in the package. Will schedule that for later in the year.