Changed incomplete tuplet formatting systemwide.
Consider this tuplet:

    >>> tuplet = Tuplet((2, 3), "c'8 d'8")

OLD:

    >>> f(tuplet)
    \times 2/3 {
            c'8
            d'8
    }

NEW:

    >>> f(tuplet)
    \tweak #'edge-height #'(0.7 . 0)
    \times 2/3 {
            c'8
            d'8
    }

The effective of the grob override is to remove the nib from the right edge of
the tuplet bracket.

The notation is standard in Sciarrino and elsewhere.

Greatly clears up examples throughout the docs.

MODEL: the rule is that any tuplet with *preprolated* duration carrying a
non-power-of-two denominator formats in this way.

It was suprising how powerfully comprehensive the existing duration model made
the implementation of a sweeping change like this.