Integrated LilyPondContextSetting into GeneralizedBeam.
NOTE: A good reason to use classes like LilyPondContextSetting and
LilyPondGrobOverride for formatting \set and \override expressions is that it
decouples Abjad's model of those expressions from LilyPond's setting syntax.
Consider the slow evolution of LilyPond's override syntax:

    \override Staff #'foo #'bar #'baz = #10
    \override Staff #'(foo bar baz) = #10
    \override Staff.foo.bar.baz = #10
    \override Staff.foo.bar.baz = 10

If all overrides pass through LilyPondGrobOverride, then there's only one place
in the system we need to update when we settle on a single version of
LilyPond's override syntax to target.