Improved trill formatting in LilyPondFormatManager.
TrillSpanner contributions to the `right` format slot are now always placed at
the end of that slot's contributions.

This is because the "pitch" which follows a `\startTrillSpan` command in a
pitched trill is really a note-event as far as LilyPond is concerned. If a
tie-event follows that note-event, the note-event will consume the tie, and no
tie grob will be printed.

By always guaranteeing that trills format farthest to the right, we prevent
ties from disappearing.