Renamed sequencetools.iterate_sequence_boustrophedon().
sequencetools.iterate_sequence_forward_and_backward_overlapping() was the old
name.

'Boustrophedon' is an (admittedly unusual) English word that means 'in the way
that oxen turn (when plowing a field)'. The word exists in the language because
this was the direction of writing for several centuries of classical Greek
(first towards the right and then back towards the left; compare with
right-to-left for Semitic languages and, later, left-to-right for languages
written with Roman characters).

The term was later picked up by researchers working the properties of integer
series.