Added three new ordinal constants: Less, More, Exact.
NEW: these should be mentioned in the 2.15 release notes in conjunction with
the sequencetools overhaul.

EXPLANATION: the constants function like Left, Right, Center but are
dimensioned for 'value' instead of for 'x' or 'y'.

The constants will be used to replace the three functions ...

    * sequencetools.partition_sequence_by_weights_at_least()
    * sequencetools.partition_sequence_by_weights_at_most()
    * sequencetools.partition_sequence_by_weights_exactly()

... with ...

    * sequencetools.partition_sequence_by_weights(..., allow_last_part=More)
    * sequencetools.partition_sequence_by_weights(..., allow_last_part=Less)
    * sequencetools.partition_sequence_by_weights(...., allow_last_part=Exact)

... instead.

The constants will also be used to replace the three functions ...

    * sequencetools.repeat_sequence_to_weight_at_least()
    * sequencetools.repeat_sequence_to_weight_at_most()
    * sequencetools.repeat_sequence_to_weight_exactly()

... with ...

    * sequencetools.repeat_sequence_to_weigth(..., allow_total=More)
    * sequencetools.repeat_sequence_to_weigth(..., allow_total=Less)
    * sequencetools.repeat_sequence_to_weigth(..., allow_total=Exact)

... instead.

There's no other intended use for the three constants at this point. Though
they are now freely available in the global namespace for other applications
that make sense.