Added Sequence to sequencetools.
The class is currently a stub.

MODEL: the class will probably eventually allow for the reimplementation of
essentially all public sequencetools functions as Sequence methods. That is not
a design goal for the 2.15 release of the system. During 2.15 development the
predicates and attribute methods in sequencetools will be reimplemented as
Sequence methods. These include:

    * sequencetools.is_monotonically_decreasing_sequence()
    * sequencetools.is_monotonically_increasing_sequence()
    * sequencetools.is_permutation()
    * sequencetools.is_repetition_free_sequence()
    * sequencetools.is_restricted_growth_function()
    * sequencetools.is_strictly_decreasing_sequence()
    * sequencetools.is_strictly_increasing_sequence()

The changes in 2.16 will change ...

    * sequencetools.flatten_sequence()
    * sequencetools.increase_elements()
    * sequencetools.permute_sequence()

... to ...

    * Sequence.flatten()
    * Sequence.increase_elements()
    * Sequence.permute()

... instead.

This will have the effect of changing sequencetools-heavy composition from
predominantly procedural to predominantly object oriented.