Cleaned up spanner-related tests system-wide.
NOTE: It's often difficult to debug tests which fail on a call to
TestManager.compare because pytest doesn't print any human-readable information
about the failure, except that some component failed to compare to a given
string:

    assert systemtools.TestManager.compare(
        staff,
        r'''
        some_string,
        '''
        )

This can be made much more useful by adding the LilyPond format of the
component under test as the second argument to the assert:

    assert systemtools.TestManager.compare(
        staff,
        r'''
        some_string,
        '''
        ), format(staff)

Should the above test fail, pytest will show both the comparison string as well
as the LilyPond format of the component. Much nicer!