Changed PYTHONPATH requirements again.
Score manager cleanup is underway.

Part of the cleanup will mean changing long import statements like ...

    from experimental.tools.scoremanagertools.managers.PackageManager \
        import PackageManager

... to shorter import statements like this ...

    from scoremanagertools.managers.PackageManager import PackageManager

... instead.

This means that Python needs to be able to the find the scoremanagertools
package.

The scoremanagertools package is housed in the experimental/tools directory.

So developers will need to add the experimental/tools directory to their
PYTHONPATH environment variable.

Something like ...

    export PYTHONPATH=$ABJAD/experimental/tools:$PYTHONPATH

... should work.

Note that py.test will fail as of this commit until the experimental/tools
directory is added to your PYTHONPATH.