Deprecated score manager MaterialPackageMaker.
Use score manager MaterialPackageManager instead.

NOTE: we're preserving the *manager* class and getting rid of the *maker*
class.

MODEL: we've engineered a very specific meaning for a "maker": the
rhythm-makers are the best example; the segment-makers in experimental are
being developed in exactly the same way as the rhythm-makers. The use of
'maker' in the now-deprecated MaterialPackageMaker class (and, indeed, in all
of the materialpackagemakers classes) really describes something else: the
UI-guided creation of a material package. This is really just another
management use case. Hence the deprecation of MaterialPackageMaker in favor of
MaterialPackageManager.