Added new score manager (E) command.
Where: available when managing segments.

Example:

    score > segments > 1 > E

Function: the (E) command opens your text editor to the top line of a segment
definition module.

Note that this works in parallel to the familiar (e) command ...

    score > segments > 1 > e

... which opens your text editor to the bottom line of a segment definition
module.

Note that this is the first uppercase command.

Note also that the (E) command is hidden.