Added StylesheetFileWrangler test.
Test ensures that menu entries are annotated with labels like (Abjad)
showing where each stylesheet is housed.

Example:

    Score manager - stylesheets

        1: clean-letter-14.ily (Abjad)
        2: clean-letter-16.ily (Abjad)
        3: rhythm-letter-16.ily (Abjad)
        4: time-signature-context.ily (Abjad)
        5: baca-letter-14.ily (library)
        6: red-example-score-stylesheet.ily (Red Example Score)
        7: styleshet.ily (L'archipel du corps)
        8: stylesheet.ily (Traiettorie inargentate)

        new (new)
        copy (cp)
        rename (ren)
        remove (rm)

    >

The label (Abjad) means the stylesheet is housed in the Abjad stylesheets
directory.

The label (library) means the stylesheet is house in the stylesheets directory
of the user's asset library.

Other stylesheets are labeled with score names.