Replaced systemtools.IOManager with score manager session.io_manager.
MODEL: an important feature of the entire score manager system is that *every*
actor in the collaboration holds a reference to a unique session variable. This
session variable in turn holds a reference to a unique score manager IOManager.
The score manager IOManager inherits from the mainline IOManager. So this means
that every actor in the collaboration can call ALL IO methods at any point.
Just dereference the session.