Reimplemented view management in terms of TypedOrderedDict.
This proves that TypedOrderedDict storage format now works as it should. This
closes GitHub #357.

This also contributes to GitHub #353 (which asks to remove 'custom_identifier'
from as many classes as possible). The TypedOrderedDict now used to persist
view inventory to disk means that View.custom_identifer is no longer used at
all.