Added 'all segments - rerender current pdf (rr)' command.
Description: (rr) iterates all segments and reinterprets any output.ly found.
The command then opens all segments' output.pdf at once.

This should answer TODO #46.

LANGUAGE: does this command operate primarily on LilyPond files or primarily on
PDFs? It's a bit tricky to say. Because the command visits (and interprets) all
LilyPond files. But the command also leaves all LilyPond files in tact. On the
other hand, the command has the effect of changing all PDFs. We could be overly
verbose in our commands and use something like ...

    all segments - rerender current pdf from current ly file (rr)

... where the from-clause makes clear both the input and the output of the
operation. But this is too much. Better will be to establish an interface
convention that determines whether our commands talk about the input to a file
transformation or about theh output of a file transformation.

The usage committed here ...

    all segments - rerender current pdf (rr)

... talks about the *output* of a file transformation. I think it would
propbably be better to talk about the *input* to the file transformation
instead. But by way of explanation, I think the verb '(re)render' suggests both
the input to a transformation and the output of a transformation; so I think
the verb is what pushed the output focus in the current commit.

A future commit will probably change this command to ...

    all segments - reinterpret current ly file (ll)

... which should be clearer all around because the verb '(re)interpret' would
seem to suggest the input to a transformation *only*.