Renamed is_secondary=False menuing keyword.
The old name was is_hidden=False.

MODEL: this commit introduces the idea of 'secondary commands'. These are the
commands revealed during any session with the (o) command. The idea is that
these commands aren't so much hidden as much as secondary to the primary use
cases exhibited at any location in the UI.