Added 'controller_stack' property to Session.
Sessions now maintain a (last-in, first-out) stack of controllers. The session
controller stack functions exactly the same way as the session breadcrumb
stack. (And, in fact, the sesssion controller stack could probably replace the
session breadcrumb stack entirely.)

LANGUAGE: this commit introduces the term 'controller' to refer to an object
that is either a wrangler or else a manager.

Using this language we can observe that the entire score management system is,
for example, implemeneted around controller delegation.