Renamed 'commands - hidden (n)'.
The old name was 'commands - navigation (n)'.

The concept of navigation commands is now removed.

MODEL: here are the menuing command classes:

    * primary
    * secondary
    * hidden

Those are the only three.

PRIMARY commands appear immediately when you interact with a controller.

SECONDARY commands appear only with the (o) command.

HIDDEN commands appear only with the (n) command.