Added selectortools.Selector to experimental.
Selector allows a way for composers to specify the manner in which components
are to be selected, to persist that selection specification, and to call it
against any valid input.

The following example demonstrates a selector configured to select the last
component in all runs of Notes longer than 1:

    >>> staff = Staff(r"c'4 \times 2/3 { d'8 r8 e'8 } r16 f'16 g'8 a'4")
    >>> selector = selectortools.Selector()
    >>> selector = selector.by_leaves()
    >>> selector = selector.by_run(Note)
    >>> selector = selector.longer_than(1)
    >>> selector = selector[-1]
    >>> for selection in selector(staff):
    ...     print selection
    ...
    ContiguousSelection(Note("d'8"),)
    ContiguousSelection(Note("a'4"),)

NOTE: All names and interfaces are provisional.