Made ScoreManagerObject 'backtracking' property private. MODEL.
MODEL: the conventions about public and private attributes in the scoremanager
branch differ somewhat from those in other parts of the system. What has
evolved to be the case is that *public attributes implement end-user-visible
functionality in the UI*. Private attributes are everything else: all the
back-end plumbing necessary to let the different components collaborate,
maintain session information, handle lower-level IO, and so on.