Added Manager._is_git_added().
These methods are now both available:

    * Manager._is_git_added()
    * Manager._is_git_versioned()

The first returns true when a file or directory has been added (or 'staged')
for commit but has not yet been committed.

The second returns true when a file or directory has been committed.

The methods make better tests possible for remove and rename operations.