Wrote import test.
Test makes sure that scoremanager.materials module is not imported with the
rest of the scoremanager package.

This matters because the scoremanager.materials module simulates user code. An
important point about user code is that user code may contain all sorts of
errors, including errors that could raise an exception to the top of the stack.
The system needs to allow for this. The system also needs to start successfully
even if user code raises exceptions. The solution is for the system never to
import user code.