Improved argument range parsing.
It is now possible to type asset names like 'clean-letter-14.ily'.

This was previously not possible. The reason was that the hyphens internal to
the file name caused the menuing system to interpret input as an argument range
(like 3-5) rather than as a candidate string for matching.

This commit attemps candidate string matching before attempting argument range
processing.