Changed 'svn st -u' to just 'svn st'. PERFORMANCE.
PERFORMANCE: this was an interesting oversight. Setting the Subversion -u flag
on 'svn st' is documented this way:

  With -u, add working revision and server out-of-date information.

This means that the -u flag causes the Subversion client to communicate with
the Subversion server, probably across the internet.

With the -u flag set, the performance of py.test battery becomes bound in part
by network performance. Removing the -u flag makes the py.test battery two or
even three times faster if (but only if) you have Subversion-managed assets on
your filesystem. (So if you're using only Git on your machine, you won't notice
any speed up in the rate at which the py.test battery runs.)