Wrote ScorePackageManager.add() test.
The test starts with an Subversion-managed score package resident in the user
score packages directory, if any is found. (If no Subversion-managed score
package is found in the user score packages directory then the test passes
silently.)

The test then makes sure it is possible to drop two unknown files into the
score package, add them to Subversion (with committing them to Subversion),
unadd them, and then leave the score package as it was found.