Wrote another ScorePackageManager.add() test.
The test starts with a Git-managed score package resident in the Abjad
score packages directory.

The test then makes sure it is possible to drop two unknown files into the
score package, add them to Git (with committing them to Git),
unadd them, and then leave the score package as it was found.