Implemented ScoreManager._get_svn_score_name().
The method takes no arguments and does what you would expect.

To be used in writing tests.

Many tests require a Subversion-managed score package. The Abjad repository is
100% Git. So this method can be used to retrieve the name of the first
(up-to-date) Subversion-managed repository in the user's score packages
directory, if any exists.

A standard pattern for testing looks like this:

    def test_ScorePackageManager_repository_status_02():
        r'''Works with Subversion.
        '''

        name = score_manager._find_svn_score_name()
        if not name:
            return

        input_ = 'ssl {} rst default q'.format(name)
        score_manager._run(pending_user_input=input_)

        string = 'Press return to continue.'
        assert score_manager._transcript.entries[-4].lines[0] == string

The input_ line will format to something like ...

    'ssl betoerung rest default q'

... which will then pass through the Score Manager on the user's system when
the tests run.