Extended ControllerContext.
Added reset_hide_hidden_commands=True keyword to ControllerContext initializer.

Set to true when you want a controller to hide the output of (?) on enter.

Set to false when you want a controller not to hide the output (?) on enter.

The pattern is probably fixed such that Menu will not hide the output of (?)
while all other controller will hide the output (?).