Swapped (<) and (>) for (<<) and (>>).
OLD: use (<) and (>) for sibling score navigation; use (<<) and (>>) for
sibling asset navigation.

OLD: use (<) and (>) for sibling asset navigation; use (<<) and (>>) for
sibling score navigation.

The rationale is that the single-key alises are faster to type than the two-key
aliases and that the quicker aliases should be assigned to more frequent use
case (ie, moving between materials or between segments; moving between scores
is less common).