Migrating AttributeDetail initializer signature. CODING STANDARDS.
OLD:

    AttributeDetail.__init__(*args)

NEW:

    AttributeDetail.__init__(
        display_string=None,
        editor_callable=None,
        is_positional=True,
        menu_key=None,
        name=None,
        retrievable_name=None,
        )

The new signature will be self-documenting.

The current commit establishes an intermediate signature to allow migration
between the old and new.

CODING STANDARDS: use of *args is basically discouraged everywhere.

The reasons for discouraging *args have to do with our documentation regime
(and Sphinx's way of working) and not with Python language features.