Removed MaterialPackageManager menu default.
OLD: defaulted to 'definition module - edit' [dme].

NEW: no default.

Explanation: the idea behind the menu default was to allow a single press of
return to execute the most frequent use case in a menu. This will still be a
good idea in some places. But this is probably a less good idea in other
places. Less good places include those menus that afford viewing PDFs and other
artifacts that launch an external application; this means frequent pressing of
return outside of the score manager and correspondingly frequent accidental
execution of the default menu command.