Implemented MaterialPackageManager.configure_autoeditor().
Also implemented MaterialPackageManager.remove_autoeditor().

MODEL: The purpose of the methods will be to configure a material package for
management by an editor class. This replaces the old idea of 'manager-made'.
Previously there was a distinction between 'handmade' and 'manager-made'
material packages. Now there are just material packages. And a material package
may be configured for autoediting or not.