Extended StylesheetWrangler (rm) command.
You can now remove multiple with range selection.

Example: Consider the following score. We want to remove the orange, purple and
white stylesheets. We can do this with a single command.

The menu prompts us to type a special confirmation string because we are
removing more than one styleshseet at one time:

    Red Example Score (2013) - stylesheets

        1: orange-stylesheet.ily
        2: purple-stylesheet.ily
        3: red-example-score-stylesheet.ily
        4: white-stylesheet.ily

        stylesheets - copy (cp)
        stylesheets - new (new)
        stylesheets - remove (rm)
        stylesheets - rename (ren)

    > rm

    Enter stylesheets to remove> 1-2, 4

    Type 'remove 3' to proceed> remove 3

    Red Example Score (2013) - stylesheets

        1: red-example-score-stylesheet.ily

        stylesheets - copy (cp)
        stylesheets - new (new)
        stylesheets - remove (rm)
        stylesheets - rename (ren)

    >

The same behavior will be ported to all wranglers.