Made Wrangler._remove() private.
MODEL: the inheritance model is that Wrangler._remove() is private and is the
workhorse method that does everything. Then the following methods ...

    * BuildFileWrangler.remove_build_file()
    * DistributionFileWrangler.remove_distribution_file()
    * MakerModuleWrangler.remove_maker_module()
    * MaterialPackageWrangler.remove_material_package()
    * ScorePackageWrangler.remove_score_package()
    * SegmentPackageWrangler.remove_segment_package()
    * StylesheetWrangler.remove_stylesheet()

... are single-statement wrappers around the private method in the baseclass.

This is true for StylesheetWrangler.remove_stylesheet() and
MakerModuleWrangler.remove_make_module() as of this commit.

The pattern will be extended to the other wranglers in coming commits.