Extended MenuSection._argument_range_string_to_numbers().
The method now knows how to detect argument_range_string input that is not
in fact an argument range. This is helpful because the method can then treat
the argument_range_string input as a sigle string against which to match.

This commit also extend Menu._argument_range_string_to_numbers() to match
against argument_range_string input in a case-neutral way. This means that ...

    Red Example Score
    RED EXAMPLE SCORE
    red example score
    rEd eXAMplE scOrE

... will all match and return the index of the score in the target menu
section.