Added count=None keyowrd to stringtools.pluralize().
Example:

    >>> stringtools.pluralize(identifier, count=0)
    'cherries'

    >>> stringtools.pluralize(identifier, count=1)
    'cherry'

    >>> stringtools.pluralize(identifier, count=2)
    'cherries'

    >>> stringtools.pluralize(identifier, count=3)
    'cherries'

    >>> stringtools.pluralize(identifier, count=99)
    'cherries'

SYNTAX: note that grammatical number in English distinguishes not between
'singular' and 'plural' but between the open-closed range (0, 1] and all other
numbers. That is, (0, 1] of a noun receives SING morphology while any other
number of nouns receive PL morphology. Specifically this means that the idea
that "more than 1 of something receives PL" is absolutely wrong: 0 of
something always recevies PL: "no cherries". (And 0 is less than 1.)

The commit here accounts for integer numbers of a noun. Set the count=None
keyword to the number concerned and pluralization will happen as it should.