Harmonizing ScoreManager and ScorePackageManager.
MODEL: it's possible that the public functionality of ScoreManager can removed
entirely in favor of the public functionality of ScorePackageManager.

The next series of commits will migrate as much functionality as possible from
ScoreManager to ScorePackageManager. The goal will be to leave ScoreManager as
only a point of delegation with all user functionality handled by a wrangler or
manager.