Implemented NullContextManager.
Instantiate like this:

    manager = systemtools.NullContextManager()

PURPOSE: use a null context manager when you need to build a with-statement
that composes two or more context managers, some of which should be defined to
do nothing.

EXAMPLE:

    directory_change = systemtools.NullContextManager()
    if self._session.is_in_score:
        path = self._get_current_directory_path()
        directory_change = systemtools.TemporaryDirectoryChange(path)
    with context, directory_change:
        pass

Note that the 'directory_change' context manager should be defined only when
the session is in score; otherwise there should effectively be no
'directory_change'. This is accomplished by first definining 'directory_change'
equal to an instance of the NullContextManager and then conditionally
redefining 'directory_change' to an instance of the TemporaryDirectoryChange
context manager.