Added 'draft' menu to BuildFileWrangler. WORKFLOW.
MODEL: the build directory is used to build both a draft score and a finished
score. There are now menu sections to handle each:

    draft latex - edit (de)
    draft latex - generate (dg)
    draft latex - typeset (dt)
    draft pdf - open (do)

    score latex - edit (se)
    score latex - generate (sg)
    score latex - typeset (st)
    score pdf - open (so)

WORKFLOW: how do the draft (score) and (finished) score differ? The draft is a
naive concatenation of segment PDFs; note that the draft is created by running
LaTeX against a .tex file that includes all segment PDFs. The (finished) score
is created by running LilyPond against a master .ly file that includes all
segment .ly files.

PERFORMANCE: how long does it take to create a draft? How long does it take to
create a (finished) score? Creating a draft is quick because a draft is only a
concatenation of preexisting PDFs. Creating a (finished) score takes more time
because LilyPond is involved.

WORKFLOW: at what point in the compositional process do composers create drafts?
At what point in the process do composers create a (finished) score? Composers
will create drafts throughout all parts of the compositional process; create a
draft anytime you want to see what your score looks like so far: this can be
done as early as when only the front and back covers exist or as late as the
final stages of composition when all segments are complete. Begin the process
of iteratively creating a (finished) score when you want to start working on
what LilyPond does at the boundaries *between sections*. This will likely come
later in the compositional process. But note that you can creat a (finished)
score at any point during the compositional process and that it might be useful
to do this to on-goinginly check LilyPond's behavior is rendering consecutive
segments.