Changed (lycp) to (mc).
Location: BuildFileManager (u).

OLD: 'segments - copy lilypond files' (lycp).

NEW: 'music - collect segment files' (mc).

The new name is use-case-centered. Composers use (mc) only to create music.tex
and music.pdf files. (The old name was explicit that copying was the underlying
action of the command; but the old name gave no indiciation of when or why the
command should be used.)

TERMINOLOGY: The new verb 'collect' conveys that files are being hunted down
from somewhere on the filesystem and copied into the local directory.

TERMINOLOGY: note that 'music' has a specific meaning in the build directory.
The word is used to refer to the (multipage) 'guts' of a score that contains
all the music (and exludes the covers and preface). It might be even more clear
(though coarser) to substitute 'guts' for 'music' in the user interface.