Implemented IOManager.get_greatest_version_number().
The method knows how to look into the versions/ directory of any segment
package or material package and pick out the greatest version number attached
to any artifact.

This is useful during the work of versioning a new set of artifacts: the
greatest existing version number must be found so that it can then be
incremented prior to persistence.