Added 'allow_item_edit' property to ListAutoeditor.
Explanation: The value is set as an initializer keyword that defaults to true.
Set ListAutoeditor.__init__(..., allow_item_edit=False, ...) to initialize a
list autoeditor that is read-only: this allow the composer to rearrange list
items (using move and remove) but will not allow the composer to edit the items
contained in the list (pressing 1, 2, 3 or so on).

The current use case is in view creation: the composer is allow to rearrange
items in the view but not to edit items in the view.